Create Jinja2 extension and tests:
Update the README with more information on how to implement the Omniture
package with Django and Tarbell applications.

Update the requirements to include Flask and Jinja and their
dependencies.

Create a Jinja2 extension, which functions much like a Django custom template
tag. This will allow users to drop in a custom tag in their Tarbell templates,
which will render an Omniture JS object or script based on the context variables
a user defines in the custom tag or in the tarbell_config.

Related to Unfuddle #295 in the Admin project.